Reworked into Jinja2 template
