+++ kmapper/template.html
@@ -1,0 +1,92 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <div id="json_graph" data-graph='{{ json_graph }}' style="display:none"></div>
+    <meta charset="utf-8">
+    <meta name="generator" content="KeplerMapper">
+    <title>{{ title }} | KeplerMapper</title>
+    <link rel="icon" type="image/png" href="https://i.imgur.com/2eZcTZn.png" />
+    <link href='https://fonts.googleapis.com/css?family=Roboto+Mono:700,300'
+          rel='stylesheet' type='text/css'>
+    <style>
+      * {margin: 0; padding: 0;}
+      html, body {height: 100%%;}
+      body {font-family: "Roboto Mono", "Helvetica", sans-serif; font-size:14px;}
+      #display {color: #95A5A6; background: #212121;}
+      #print {color: #000; background: #FFF;}
+      h1 {font-size: 21px; font-weight: 300; font-weight: 300;}
+      h2 {font-size: 18px; padding-bottom: 20px; font-weight: 300;}
+      h3 {font-size: 14px; font-weight: 700; text-transform: uppercase;}
+      #meta h3 { float: left; padding-right: 8px;}
+      p, #tooltip h3, ol, ul, table {padding-bottom: 20px;}
+      ol, ul {padding-left: 20px;}
+      ol b {display: block;}
+      a {color: #16a085; text-decoration: none;}
+      a:hover {color: #2ecc71;}
+      #header {height: 35px; padding: 20px; position: absolute; top: 0; left: 0; right: 0;
+               z-index: 9999;}
+      #display #header {background: #111111; box-shadow: 0px 0px 4px #000}
+      #print #header {background: #FFF;}
+      #canvas {height: 100%%; width: 100%%; display: block;}
+      #tooltip {position: absolute; top: 75px; left: 0; bottom: 0;
+                width: 320px; padding: 20px; overflow: auto; display: none;}
+      #display #tooltip {background: #191919;}
+      #print #tooltip {background: #FFF;}
+      #meta {position: absolute; top: 75px; right: 0; bottom: 0;
+             width: 320px; padding: 20px; overflow: auto;}
+      #display #meta {background: #191919;}
+      #print #meta {background: #FFF; }
+      #meta_control, #tooltip_control {position: absolute; right: 20px;}
+      #meta::-webkit-scrollbar, #tooltip::-webkit-scrollbar {width: 1em;}
+      #meta::-webkit-scrollbar-track, #tooltip::-webkit-scrollbar-track {
+        -webkit-box-shadow: inset 0 0 6px rgba(0,0,0,0.3);}
+      #meta::-webkit-scrollbar-thumb, #tooltip::-webkit-scrollbar-thumb {
+        background-color: darkgrey; outline: 1px solid slategrey;}
+      #histogram { display: block; height: 100px; padding-top: 50px; clear: both;}
+      #display #histogram {opacity: 0.68;}
+      .bin {width: 30px; float: left;}
+      .bin div { font-size: 10px; display: block; width: 35px; margin-top: -30px;
+                 text-align: right; margin-left:-3px;
+                 -webkit-transform: rotate(-90deg); -moz-transform: rotate(-90deg);
+                 -ms-transform: rotate(-90deg); -o-transform: rotate(-90deg);}
+      #histogram:hover {opacity:1.;}
+      #display .circle {stroke-opacity:0.18; stroke-width: 7px; stroke: #000;}
+      #print .circle {stroke-opacity:1; stroke-width: 2px; stroke: #000;
+                      stroke-linecap: round;}
+      #print .link {stroke: #000;}
+      #display .link {stroke: rgba(160,160,160, 0.5);}
+      table { border-collapse: collapse; display: table; width: 100%%; margin-bottom:20px;}
+      td, th { padding: 5px; text-align: left;}
+      #display th { background: #212121}
+      td { border-bottom: 1px solid #111;}
+    </style>
+  </head>
+  <body id="display">
+    <div id="header">
+      <noscript><b>Requires JavaScript (d3.js) for visualizations</b></noscript>
+      <h1>{{ title }}</h1>
+    </div>
+    <div id="canvas">
+
+    </div>
+    <div id="tooltip">
+      <div id="tooltip_control">
+        <a href="#"><small>[-]</small></a>
+      </div>
+      <h2>Cluster Meta</h2>
+      <div id="tooltip_content">
+
+      </div>
+    </div>
+    <div id="meta">
+      <div id="meta_control">
+        <a href="#"><small>[-]</small></a>
+      </div>
+      <h2>Graph Meta</h2>
+      {{ meta }}
+      {{ color_distribution }}
+    </div>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/d3/3.5.5/d3.min.js"></script>
+    <script src='{{ script_path }}'></script>
+  </body>
+</html>
